PR-C1j: Route content_pipeline temporal through reasoning core
Replaces extracted_content_pipeline/reasoning/temporal.py (~466-line
drifted fork) with a thin wrapper that re-exports the canonical
TemporalEngine + 5 dataclasses + constants from
extracted_reasoning_core (consolidated via PR-C1b / PR #100, with
the rich type promotion landing in PR-C1c / PR #102).

Surface comparison suggested PR-C1j would be a clean wrapper
conversion since all the public types and constants were already in
core. Tests revealed the slim core's TemporalEngine had several
latent bugs that PR-C1c's frozen-dataclass promotion activated --
atlas production uses its own atlas_brain/reasoning/temporal.py and
no test exercised core's analyze_vendor path, so the breakage went
unnoticed. PR-C1j fixes them in core (since core is the canonical
home) and routes content_pipeline through the now-working core:

  1. analyze_vendor mutated frozen TemporalEvidence -- triggered
     FrozenInstanceError on the `evidence.velocities = ...` line.
     Now constructs TemporalEvidence once with all fields. Also
     normalizes vendor_name via str(...).strip() so loosely-shaped
     callers don't break the per-vendor lookup.
  2. _compute_long_term_trends mutated frozen LongTermTrend, and
     blew up on string-shaped snapshot_date. Now uses _coerce_date
     for date math and constructs LongTermTrend with final fields.
  3. _compute_percentiles depended on atlas_brain helpers that the
     extracted standalone CI doesn't ship -- replaced with the
     self-contained SELECT content_pipeline had been using. Atlas
     parity preserved via _percentiles_from_rows(min_count=...).
  4. _compute_velocities and _compute_anomalies now use _numeric_value
     so messy snapshot strings ("1,234", "28%") coerce cleanly
     instead of dropping through bare float() ValueError handlers.
  5. analyze_all_vendors normalizes vendor_name + degrades safely
     when the pool doesn't expose fetch().
  6. _infer_category dropped (atlas-coupled, no longer called -- the
     percentile path now derives category directly from the latest
     snapshot's product_category column).
  7. Module-level helpers added: _coerce_date, _days_between,
     _volatility, _percentiles_from_rows -- carried from
     content_pipeline's pre-wrapper fork. _percentiles_from_rows
     accepts min_count to preserve TemporalEngine's
     min_days_for_percentiles parameterization (PR-C1b's contribution).

Test results: 27/27 temporal tests green (including 5 that previously
exercised the broken paths). 382/382 broader extracted-reasoning +
downstream temporal consumer tests green. 331/331 standalone
extracted-pipeline CI green locally. Net: -381 LOC across the two
temporal files (content_pipeline 466 -> 49 wrapper; core 556 -> 592,
slight grow for module-level helpers + _percentiles_from_rows, more
than offset by content_pipeline shrink).
